[Rename]: 폴더 구조 변경 - (admin) 하위로 이동
