+++ frontend/app/companies/[id]/(admin)/layout.tsx
@@ -1,0 +1,10 @@
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      <div className="flex h-lvh w-full">
+        <div className="w-[200px] h-full bg-gray-200">sidebar</div>
+        <div className="flex justify-center w-full relative">{children}</div>
+      </div>
+    </>
+  );
+}
